\name{rg_graph_stubs}
\alias{rg_build_kg}
\alias{rg_query_kg}
\title{Knowledge graph stubs}
\usage{
rg_build_kg(...)
rg_query_kg(...)
}
\arguments{
\item{...}{Reserved for future knowledge graph support.}
}
\description{
Stable stubs for future GraphRAG or knowledge graph support.
}
\value{Raises a clear not-implemented error.}
