\name{rg_integrated_stubs}
\alias{rg_init_integrated_project}
\alias{rg_draft_integrated_guide}
\title{Integrated guide stubs}
\usage{
rg_init_integrated_project(...)
rg_draft_integrated_guide(...)
}
\arguments{
\item{...}{Reserved for future integrated guide support.}
}
\description{
Stable stubs for future iADRG/icSDRG support.
}
\value{Raises a clear not-implemented error.}
